fixed add() and subtract() function to handle float input values correctly

+++ calculator.js
@@ -1,7 +1,7 @@
 function add(a, b){
     
-    return parseInt(a) + parseInt(b);
+    return parseFloat(a) + parseFloat(b);
 }
 function subtract(a,b){
-    return parseInt(a) - parseInt(b);
+    return parseFloat(a) - parseFloat(b);
 }
